refactor: create useGenerateRecoveryQrAndPoll hook for RecoveryCodeScreen
