+++ BrightID/src/components/Onboarding/RecoveryFlow/RecoveryCodeScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import {
   Alert,
   StatusBar,
@@ -9,8 +9,6 @@ import {
 } from 'react-native';
 import Clipboard from '@react-native-community/clipboard';
 import Svg, { Path } from 'react-native-svg';
-import qrcode from 'qrcode';
-import { parseString } from 'xml2js';
 import { path } from 'ramda';
 import Spinner from 'react-native-spinkit';
 import { useTranslation } from 'react-i18next';
@@ -18,7 +16,7 @@ import { useFocusEffect, useNavigation } from '@react-navigation/native';
 import i18next from 'i18next';
 import CheckBox from '@react-native-community/checkbox';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import { useDispatch, useSelector } from '@/store/hooks';
+import { useDispatch } from '@/store/hooks';
 import {
   BLACK,
   DARK_PRIMARY,
@@ -32,28 +30,12 @@ import {
 import { fontSize } from '@/theme/fonts';
 import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import { RecoveryErrorType } from './RecoveryError';
-import { setupRecovery } from './thunks/recoveryThunks';
-import { buildRecoveryChannelQrUrl } from '@/utils/recovery';
-import {
-  createRecoveryChannel,
-  pollRecoveryChannel,
-} from './thunks/channelThunks';
-import {
-  resetRecoveryData,
-  selectRecoveryStep,
-  setRecoverStep,
-  uploadCompletedByOtherSide,
-} from './recoveryDataSlice';
-import {
-  clearImportChannel,
-  createSyncChannel,
-  pollImportChannel,
-  setupSync,
-} from '../ImportFlow/thunks/channelThunks';
+import { resetRecoveryData, setRecoverStep } from './recoveryDataSlice';
+import { clearImportChannel } from '../ImportFlow/thunks/channelThunks';
 import { recover_steps, UNIVERSAL_LINK_PREFIX } from '@/utils/constants';
-import { userSelector } from '@/reducer/userSlice';
 import BrightIDLogo from '@/components/Icons/BrightIDLogo';
 import Copy from '@/components/Icons/Copy';
+import { useGenerateRecoveryQrAndPoll } from './useGenerateRecoveryQrAndPoll';
 
 /**
  * Recovery Code screen of BrightID
@@ -62,119 +44,22 @@ import Copy from '@/components/Icons/Copy';
  */
 
 const RecoveryCodeScreen = ({ route }) => {
-  const { action, urlType } = route.params;
-  const [qrUrl, setQrUrl] = useState<URL>();
-  const [qrsvg, setQrsvg] = useState('');
-  const [alreadyNotified, setAlreadyNotified] = useState(false);
-  const recoveryData = useSelector((state) => state.recoveryData);
-  const { id } = useSelector(userSelector);
-  const isScanned = useSelector(
-    (state) =>
-      uploadCompletedByOtherSide(state) ||
-      state.recoveryData.recoveredConnections ||
-      state.recoveryData.recoveredGroups ||
-      state.recoveryData.recoveredBlindSigs,
-  );
   const { t } = useTranslation();
-  const dispatch = useDispatch();
   const navigation = useNavigation<NativeStackNavigationProp<any>>();
-  const step = useSelector(selectRecoveryStep);
+  const dispatch = useDispatch();
+  const { action, urlType } = route.params;
 
-  const sigCount = recoveryData.sigs
-    ? Object.values(recoveryData.sigs).length
-    : 0;
-
-  // start polling recovery channel to get sig and mutual info
-  useEffect(() => {
-    if (
-      action === 'recovery' &&
-      recoveryData.recoverStep === recover_steps.POLLING_SIGS &&
-      !recoveryData.channel.pollTimerId
-    ) {
-      dispatch(pollRecoveryChannel());
-    }
-  }, [
-    action,
-    dispatch,
-    recoveryData.channel.pollTimerId,
-    recoveryData.recoverStep,
-  ]);
-
-  // create recovery data and start polling channel
-  useEffect(() => {
-    const runRecoveryEffect = async () => {
-      // create publicKey, secretKey, aesKey for user
-      await dispatch(setupRecovery());
-      // create channel and upload new publicKey to get signed by the scanner
-      await dispatch(createRecoveryChannel());
-      dispatch(setRecoverStep(recover_steps.POLLING_SIGS));
-    };
-    const runImportEffect = async () => {
-      // create publicKey, secretKey, aesKey for user
-      await dispatch(setupRecovery());
-      // create channel and upload new publicKey to be added as a new signing key by the scanner
-      await dispatch(createRecoveryChannel());
-      // start polling channel to get connections/groups/blindsigs info
-      dispatch(pollImportChannel());
-    };
-    const runSyncEffect = async () => {
-      // create a new aesKey
-      await dispatch(setupSync());
-      // create channel and upload lastSyncTime to the channel if it is not primary device
-      // or poll lastSyncTime from other side if it is and then upload connections/groups/blindsigs
-      // added after lastSyncTime to the channel
-      await dispatch(createSyncChannel());
-      // start polling channel to get new connections/groups/blindsigs info
-      dispatch(pollImportChannel());
-    };
-
-    if (step === recover_steps.NOT_STARTED) {
-      if (action === 'recovery') {
-        if (!id) {
-          console.log(`initializing recovery process`);
-          runRecoveryEffect();
-        } else {
-          console.log(`Not starting recovery process, user has id!`);
-        }
-      } else if (action === 'import') {
-        console.log(`initializing import process`);
-        runImportEffect();
-      } else if (action === 'sync') {
-        console.log(`initializing sync process`);
-        runSyncEffect();
-      }
-    }
-  }, [action, dispatch, id, step]);
-
-  const [changePrimaryDevice, setChangePrimaryDevice] = useState(true);
-  // set QRCode and SVG
-  useEffect(() => {
-    if (recoveryData.channel.url && recoveryData.aesKey) {
-      const newQrUrl = buildRecoveryChannelQrUrl({
-        aesKey: recoveryData.aesKey,
-        url: recoveryData.channel.url,
-        t: urlType,
-        changePrimaryDevice,
-      });
-      console.log(`new qrCode url: ${newQrUrl.href}`);
-      setQrUrl(newQrUrl);
-
-      const parseQrString = (err, qrsvg) => {
-        if (err) return console.log(err);
-        setQrsvg(qrsvg);
-      };
-
-      qrcode.toString(newQrUrl.href, (err, qr) => {
-        if (err) return console.log(err);
-        parseString(qr, parseQrString);
-      });
-    }
-  }, [
+  const {
+    isScanned,
+    qrUrl,
+    qrsvg,
+    sigCount,
     changePrimaryDevice,
-    recoveryData.aesKey,
-    recoveryData.channel.url,
-    urlType,
-  ]);
+    setChangePrimaryDevice,
+    recoveryData,
+    alreadyNotified,
+    setAlreadyNotified,
+  } = useGenerateRecoveryQrAndPoll({ action, urlType });
 
   // track errors
   useEffect(() => {
@@ -240,6 +125,7 @@ const RecoveryCodeScreen = ({ route }) => {
       sigCount,
       isScanned,
       t,
+      setAlreadyNotified,
       dispatch,
       navigation,
       changePrimaryDevice,
@@ -354,7 +240,6 @@ const RecoveryCodeScreen = ({ route }) => {
             {action === 'import' && (
               <View style={styles.changePrimaryDeviceSwitchContainer}>
                 <CheckBox
-                  
                   tintColors={{ false: GREY, true: ORANGE }}
                   // style={{borderRadius: }}
                   onValueChange={(value) => {
